Seperate network_visualiser from visualiser
